+++ requests_cache/session.py
@@ -251,6 +251,8 @@ class CacheMixin(MIXIN_BASE):
             response = cached_response  # type: ignore
         elif actions.resend_request:
             response = self._resend(request, actions, cached_response, **kwargs)  # type: ignore
+            if isinstance(response, OriginalResponse) and actions._vary_matched:
+                response.update_content_changed(cached_response)
         elif actions.send_request:
             response = self._send_and_cache(request, actions, cached_response, **kwargs)
         else:
@@ -282,7 +284,10 @@ class CacheMixin(MIXIN_BASE):
         if not actions.skip_write:
             self.cache.save_response(response, actions.cache_key, actions.expires)
         elif cached_response is not None and response.status_code == 304:
-            cached_response = actions.update_revalidated_response(response, cached_response)
+            cached_response = actions.update_revalidated_response(
+                response, CachedResponse.from_response(cached_response)
+            )
+            cached_response.cache_key = actions.cache_key
             if not actions.skip_write:
                 self.cache.save_response(cached_response, actions.cache_key, actions.expires)  # type: ignore[unreachable]
             return cached_response